+++ SVGRenderer/Rect.cpp
@@ -16,9 +16,9 @@ void RectangleSVG::buildRect(vector<char*> name, vector<char*> value) {
     for (int i = 0; i < name.size(); ++i) {
         temp = name[i];
         if (temp == "x")
-            this->coordinate.setX(atoi(value[i]));
+            this->coordinate.setX(stof(value[i]));
         else if (temp == "y")
-            this->coordinate.setY(atoi(value[i]));
+            this->coordinate.setY(stof(value[i]));
         else if (temp == "width")
             this->setWidth(atoi(value[i]));
         else if (temp == "height")
